requirements.txt
pysimplegui==4.50.0
mysql-connector-python==8.0.26
matplotlib==3.4.3
pyinstaller==4.7
mplcursors==0.5.1